@@ -32,19 +32,19 @@
   "author": "kintone",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.13.10",
-    "@babel/preset-env": "^7.13.12",
+    "@babel/core": "^7.13.16",
+    "@babel/preset-env": "^7.13.15",
     "@babel/preset-typescript": "^7.13.0",
-    "@cybozu/eslint-config": "^12.0.1",
-    "@types/jest": "^26.0.21",
-    "@types/node": "^13.13.47",
+    "@cybozu/eslint-config": "^13.0.0",
+    "@types/jest": "^26.0.23",
+    "@types/node": "^13.13.51",
     "comment-json": "^4.1.0",
-    "eslint": "^7.22.0",
+    "eslint": "^7.25.0",
     "jest": "^26.6.3",
     "lerna": "^4.0.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
-    "typescript": "^4.2.3"
+    "typescript": "^4.2.4"
   }
 }
