@@ -1,6 +1,6 @@
 {
   "name": "@kintone/webpack-plugin-kintone-plugin",
-  "version": "4.0.17",
+  "version": "4.0.22",
   "description": "A webpack plugin to generate a plugin zip",
   "main": "dist/index.js",
   "engines": {
@@ -37,22 +37,22 @@
   },
   "homepage": "https://github.com/kintone/js-sdk/tree/master/packages/webpack-plugin-kintone-plugin#readme",
   "devDependencies": {
-    "@types/adm-zip": "^0.4.33",
+    "@types/adm-zip": "^0.4.34",
     "@types/chokidar": "^1.7.5",
     "@types/lodash": "^4.14.168",
     "@types/mkdirp": "^1.0.1",
     "@types/rimraf": "^3.0.0",
-    "adm-zip": "^0.5.4",
+    "adm-zip": "^0.5.5",
     "rimraf": "^3.0.2",
     "ts-node": "^9.1.1",
-    "webpack": "^5.27.2",
-    "webpack-cli": "4.5.0"
+    "webpack": "^5.35.1",
+    "webpack-cli": "4.6.0"
   },
   "peerDependencies": {
     "webpack": "^4.0.0 || ^5.0.0"
   },
   "dependencies": {
-    "@kintone/plugin-packer": "^4.0.2",
+    "@kintone/plugin-packer": "^4.0.6",
     "chokidar": "^3.5.1",
     "lodash": "^4.17.21",
     "mkdirp": "^1.0.4"
