@@ -3,17 +3,19 @@ import path from "path";
 import { KintoneRestAPIClient } from "@kintone/rest-api-client";
 import { AppID, Record } from "@kintone/rest-api-client/lib/client/types";
 import { buildRestAPIClient, RestAPIClientOptions } from "../api";
-import { KintoneRecord } from "../types";
+import { KintoneRecordForResponse } from "../types";
 import { printAsJson } from "../printers/printAsJson";
 import { printAsCsv } from "../printers/printAsCsv";
 
 export type Options = {
   app: AppID;
   attachmentDir?: string;
-  format?: "json" | "csv";
+  format?: ExportFileFormat;
   query?: string;
   exportFields?: string;
 };
+
+export type ExportFileFormat = "json" | "csv";
 
 type FileInfo = {
   name: string;
@@ -89,7 +91,7 @@ async function printRecords({
   argv,
   apiClient,
 }: {
-  records: KintoneRecord[];
+  records: KintoneRecordForResponse[];
   argv: RestAPIClientOptions & Options;
   apiClient: KintoneRestAPIClient;
 }) {
