@@ -1,10 +1,10 @@
 import { convertKintoneRecordsToCsv } from "../convertKintoneRecordsToCsv";
-import { FieldsJson, KintoneRecord } from "../../../types";
+import { FieldsJson, KintoneRecordForResponse } from "../../../types";
 
-const records: KintoneRecord[] = require("./fixtures/input.json");
+const records: KintoneRecordForResponse[] = require("./fixtures/input.json");
 const fieldsJson: FieldsJson = require("./fixtures/fields.json");
-const subTableRecords: KintoneRecord[] = require("./fixtures/subtable_input.json");
-const subTableFieldsJson: FieldsJson = require("./fixtures/subtable_fields.json");
+const subtableRecords: KintoneRecordForResponse[] = require("./fixtures/subtable_input.json");
+const subtableFieldsJson: FieldsJson = require("./fixtures/subtable_fields.json");
 
 describe("convertKintoneRecordsToCsv", () => {
   it("should convert kintone records to csv string correctly", () => {
@@ -14,12 +14,15 @@ line
 text","2021-02-10T06:14:00Z","""sample2""","16","""sample3""
 sample4"
 `;
-    expect(convertKintoneRecordsToCsv({ records, fieldsJson })).toBe(
-      expectedCsv
-    );
+    expect(
+      convertKintoneRecordsToCsv({
+        records,
+        fieldProperties: fieldsJson.properties,
+      })
+    ).toBe(expectedCsv);
   });
   it("should convert kintone records to csv string correctly when SUBTABLE included", () => {
-    const expectedCsv = `"*","recordNumber","updatedTime","dropDown","creator","subTable.id","subTable.subTableText","subTable.subTableCheckbox","modifier","richText","singleLineText","number","radioButton","multiLineText","createdTime","checkBox","calc","multiSelect"
+    const expectedCsv = `"*","recordNumber","updatedTime","dropDown","creator","subTable","subTableText","subTableCheckbox","modifier","richText","singleLineText","number","radioButton","multiLineText","createdTime","checkBox","calc","multiSelect"
 *,"9","2021-02-16T02:43:00Z","sample1","username","537306","text_line1","st_sample1","username","<div><div>rich text editor<br /></div></div><div>rich text editor<br /></div><div>rich text editor<br /></div>","""single line text""","8","sample2","multi
 line
 text","2021-02-10T06:14:00Z","""sample2""","16","""sample3""
@@ -31,8 +34,8 @@ sample4"
 `;
     expect(
       convertKintoneRecordsToCsv({
-        records: subTableRecords,
-        fieldsJson: subTableFieldsJson,
+        records: subtableRecords,
+        fieldProperties: subtableFieldsJson.properties,
       })
     ).toBe(expectedCsv);
   });
@@ -44,7 +47,7 @@ sample4"
     expect(
       convertKintoneRecordsToCsv({
         records,
-        fieldsJson,
+        fieldProperties: fieldsJson.properties,
         exportFields: "recordNumber,number",
       })
     ).toBe(expectedCsv);
