@@ -1,18 +1,18 @@
 import { convertKintoneRecordsToCsv } from "./convertKintoneRecordsToCsv";
-import { FieldsJson, KintoneRecord } from "../../types";
+import { FieldsJson, KintoneRecordForResponse } from "../../types";
 
 export const printAsCsv = ({
   records,
   fieldsJson,
   exportFields,
 }: {
-  records: KintoneRecord[];
+  records: KintoneRecordForResponse[];
   fieldsJson: FieldsJson;
   exportFields?: string;
 }) => {
   const csv = convertKintoneRecordsToCsv({
     records,
-    fieldsJson,
+    fieldProperties: fieldsJson.properties,
     exportFields,
   });
   console.log(csv);
