@@ -3,9 +3,17 @@ import {
   KintoneRecordField,
 } from "@kintone/rest-api-client";
 
-export type CsvRecords = Array<Record<string, string>>;
-export type ParsedRecord = Record<string, Record<"value", unknown>>;
-export type KintoneRecord = { [fieldCode: string]: KintoneRecordField.OneOf };
+export type CsvRows = Array<Record<string, string>>;
+export type KintoneRecordForParameter = {
+  [fieldCode: string]: {
+    value: unknown;
+  };
+};
+
+export type KintoneRecordForResponse = {
+  [fieldCode: string]: KintoneRecordField.OneOf;
+};
+export type FieldProperties = Record<string, KintoneFormFieldProperty.OneOf>;
 export type FieldsJson = {
-  properties: { [k: string]: KintoneFormFieldProperty.OneOf };
+  properties: FieldProperties;
 };
