@@ -6,7 +6,7 @@ type PlatformDeps = {
   getRequestToken: () => Promise<string>;
   getDefaultAuth: () => DiscriminatedAuth;
   buildPlatformDependentConfig: (params: object) => object;
-  buildHeaders: () => Record<string, string>;
+  buildHeaders: (params: { userAgent?: string }) => Record<string, string>;
   buildFormDataValue: (data: unknown) => unknown;
   buildBaseUrl: (baseUrl?: string) => string;
   getVersion: () => string;
